Fix link to `example.html`
When `make html` is executed, the `example.html` file is not copied to
the `_build` directory. For a reST+Sphinx project, two solutions spring
to mind:

1.  Reference the downloadable file with the `:download:` role.
2.  Place `example.html` into the `_static` directory.

In either case, Sphinx will copy `example.html` to the `_build`
directory.

The former option isn't available to us, because the file which is
referencing `example.html` is a markdown file, not a reST file. That
leaves the latter solution.